pyproject.toml
[project]
name = "pageindex-mcp"
version = "0.1.0"
description = "PageIndex MCP server - file watcher + search"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.101.0",
    "pymupdf>=1.26.4",
    "PyPDF2>=3.0.1",
    "python-dotenv>=1.1.0",
    "tiktoken>=0.11.0",
    "pyyaml>=6.0.2",
    "mcp>=1.26.0",
    "watchdog>=6.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
]
